Update to the latest version of checkstyle. (#6)
### Motivation

The dependabots update of the checkstyle version ( #5 ) failed due to changes in the way the config file needs to be structured.  This PR updates the checkstyle version, AND fixes the config file.

### Modifications

This commit updates the checkstyle version, the checkstyle maven plugin
version.

This also updates the checkstyle config file to account for changes to
checkstyle configuration since the old version.